+++ NEWS.md
@@ -1,3 +1,6 @@
+# onpoint 1.1
+* Improved function names
+
 # onpoint 1.0.6
 * Order of stable `i` and randomized pattern `j` changed in `simulate_antecedent_conditions`
 * Fix anchors in documentation